--- src/adc_readout/integration_test/ContinousSamplingTimer.cpp
@@ -1,9 +1,0 @@
-/* Copyright (C) 2017-2018 European Spallation Source, ERIC. See LICENSE file */
-//===----------------------------------------------------------------------===//
-///
-/// \file
-///
-//===----------------------------------------------------------------------===//
-
-#include "ContinousSamplingTimer.h"
-
