--- src/adc_readout/integration_test/PoissonDelay.cpp
@@ -1,8 +1,0 @@
-/* Copyright (C) 2017-2018 European Spallation Source, ERIC. See LICENSE file */
-//===----------------------------------------------------------------------===//
-///
-/// \file
-///
-//===----------------------------------------------------------------------===//
-
-#include "PoissonDelay.h"
